convert launch screen from activity to fragment, move it to separate package, plug into nav_graph instead of main screen placeholder
