+++ app/src/main/res/layout/activity_launch_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LaunchScreenActivity"
+    tools:context="com.safetyheads.akademiaandroida.launchScreen.LaunchScreenFragment"
     android:id="@+id/parent_layout"
     android:background="@color/n_0">
