+++ app/src/main/res/layout/fragment_mainscreen_placeholder.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout3"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainscreenPlaceholder">
+    tools:context=".launchScreen.LaunchScreenFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
